Create base module-details-to-module transformer
