+++ magicmetamodel/src/main/kotlin/org/jetbrains/magicmetamodel/impl/workspacemodel/impl/updaters/transformers/ModuleDetailsToJavaModuleTransformer.kt
@@ -6,21 +6,19 @@ import ch.epfl.scala.bsp4j.JvmBuildTarget
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import org.jetbrains.magicmetamodel.impl.workspacemodel.ModuleDetails
-import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.ContentRoot
 import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.JavaModule
 import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.JvmJdkInfo
-import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.Module
 import java.net.URI
 import java.nio.file.Path
 import kotlin.io.path.toPath
 
-internal object ModuleDetailsToJavaModuleTransformer : WorkspaceModelEntityTransformer<ModuleDetails, JavaModule> {
+internal object ModuleDetailsToJavaModuleTransformer : ModuleDetailsToModuleTransformer<JavaModule>() {
 
-  private const val type = "JAVA_MODULE"
+  override val type = "JAVA_MODULE"
 
   override fun transform(inputEntity: ModuleDetails): JavaModule =
     JavaModule(
-      module = toModule(inputEntity),
+      module = toModule(inputEntity, includeJavacOpts = true, includePythonOpts = false),
       baseDirContentRoot = toBaseDirContentRoot(inputEntity),
       sourceRoots = SourcesItemToJavaSourceRootTransformer.transform(inputEntity.sources.map {
         BuildTargetAndSourceItem(
@@ -37,25 +35,6 @@ internal object ModuleDetailsToJavaModuleTransformer : WorkspaceModelEntityTrans
       }),
       compilerOutput = toCompilerOutput(inputEntity),
       jvmJdkInfo = toJdkInfo(inputEntity)
-    )
-
-  private fun toModule(inputEntity: ModuleDetails): Module {
-    val bspModuleDetails = BspModuleDetails(
-      target = inputEntity.target,
-      allTargetsIds = inputEntity.allTargetsIds,
-      dependencySources = inputEntity.dependenciesSources,
-      type = type,
-      javacOptions = inputEntity.javacOptions,
-      pythonOptions = null,
-    )
-
-    return BspModuleDetailsToModuleTransformer.transform(bspModuleDetails)
-  }
-
-  private fun toBaseDirContentRoot(inputEntity: ModuleDetails): ContentRoot =
-    ContentRoot(
-      // TODO what if null?
-      url = URI.create(inputEntity.target.baseDirectory ?: "file:///todo").toPath()
     )
 
   private fun toCompilerOutput(inputEntity: ModuleDetails): Path? =
