+++ magicmetamodel/src/main/kotlin/org/jetbrains/magicmetamodel/impl/workspacemodel/impl/updaters/transformers/ModuleDetailsToPythonModuleTransformer.kt
@@ -6,21 +6,17 @@ import ch.epfl.scala.bsp4j.PythonBuildTarget
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import org.jetbrains.magicmetamodel.impl.workspacemodel.ModuleDetails
-import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.ContentRoot
-import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.Module
 import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.PythonModule
 import org.jetbrains.magicmetamodel.impl.workspacemodel.impl.updaters.PythonSdkInfo
-import java.net.URI
 import kotlin.io.path.Path
-import kotlin.io.path.toPath
 
-internal object ModuleDetailsToPythonModuleTransformer : WorkspaceModelEntityTransformer<ModuleDetails, PythonModule> {
+internal object ModuleDetailsToPythonModuleTransformer : ModuleDetailsToModuleTransformer<PythonModule>() {
 
-  private const val type = "PYTHON_MODULE"
+  override val type = "PYTHON_MODULE"
 
   override fun transform(inputEntity: ModuleDetails): PythonModule =
     PythonModule(
-      module = toModule(inputEntity),
+      module = toModule(inputEntity, includeJavacOpts = false, includePythonOpts = true),
       baseDirContentRoot = toBaseDirContentRoot(inputEntity),
       sourceRoots = SourcesItemToPythonSourceRootTransformer.transform(inputEntity.sources.map {
         BuildTargetAndSourceItem(
@@ -31,25 +27,6 @@ internal object ModuleDetailsToPythonModuleTransformer : WorkspaceModelEntityTra
       resourceRoots = ResourcesItemToPythonResourceRootTransformer.transform(inputEntity.resources),
       libraries = emptyList(),
       sdkInfo = toSdkInfo(inputEntity)
-    )
-
-  private fun toModule(inputEntity: ModuleDetails): Module {
-    val bspModuleDetails = BspModuleDetails(
-      target = inputEntity.target,
-      allTargetsIds = inputEntity.allTargetsIds,
-      dependencySources = inputEntity.dependenciesSources,
-      type = type,
-      javacOptions = null,
-      pythonOptions = inputEntity.pythonOptions,
-    )
-
-    return BspModuleDetailsToModuleTransformer.transform(bspModuleDetails)
-  }
-
-  private fun toBaseDirContentRoot(inputEntity: ModuleDetails): ContentRoot =
-    ContentRoot(
-      // TODO what if null?
-      url = URI.create(inputEntity.target.baseDirectory ?: "file:///todo").toPath()
     )
 
   private fun toSdkInfo(inputEntity: ModuleDetails): PythonSdkInfo? {
